requirements.txt
# Core dependencies
chromadb>=0.4.0
sentence-transformers>=2.2.2
numpy>=1.24.0
scikit-learn>=1.3.0

# NLP and text processing
nltk>=3.8.1

# Web framework
streamlit>=1.28.0
fastapi>=0.115.9
uvicorn>=0.34.1

# Utilities
pandas>=2.1.0
requests>=2.31.0
tqdm>=4.66.0
python-dotenv>=1.1.0
matplotlib.pyplot>=3.8.0
seaborn>=0.12.2

# Optional: for development
pytest>=7.4.0
black>=23.9.0
flake8>=6.1.0

# Normalizer
geopy

# AI tools
google-genai
